Refactor v2 trace pipe dispatch
